# packages/protocol/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: brigames-station API
  version: 0.0.0
  description: HTTP contract for the brigames-station control plane.
servers:
  - url: /
    description: API base path
paths:
  /health:
    get:
      summary: Check backend liveness
      description: Confirms that the HTTP backend process is alive. This endpoint does not depend on PostgreSQL.
      operationId: getHealth
      security: []
      responses:
        '200':
          description: Backend process is alive.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LivenessStatus'
        '500':
          description: Unexpected backend failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStatus'
  /ready:
    get:
      summary: Check backend readiness
      description: Confirms that the backend can reach PostgreSQL and is ready to serve requests.
      operationId: getReadiness
      security: []
      responses:
        '200':
          description: Backend is ready.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessStatus'
        '503':
          description: Backend is not ready because PostgreSQL is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadinessStatus'
  /auth/register:
    post:
      summary: Register a member account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/RegistrationDisabled'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      summary: Authenticate with username or email
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      summary: Rotate a refresh token
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens rotated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: Revoke a refresh token
      operationId: logout
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Refresh token revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      summary: Get the authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      summary: Update the authenticated user's avatar
      operationId: updateCurrentUserAvatar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAvatarRequest'
      responses:
        '200':
          description: Avatar updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /servers:
    get:
      summary: List the authenticated user's servers
      operationId: listServers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Servers in which the current user is a member.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a server
      description: Creates a server, an owner membership for the caller, and its `general` text channel atomically.
      operationId: createServer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServerRequest'
      responses:
        '201':
          description: Server created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /servers/{serverId}:
    parameters:
      - $ref: '#/components/parameters/ServerID'
    get:
      summary: Get a server visible to the authenticated user
      operationId: getServer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Server details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ServerNotFound'
  /servers/{serverId}/channels:
    parameters:
      - $ref: '#/components/parameters/ServerID'
    get:
      summary: List a server's text channels
      operationId: listChannels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Channels visible to the current server member.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ServerNotFound'
    post:
      summary: Create a text channel
      operationId: createChannel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Text channel created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ServerNotFound'
  /servers/{serverId}/leave:
    parameters:
      - $ref: '#/components/parameters/ServerID'
    post:
      summary: Leave a server
      operationId: leaveServer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Membership removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ServerNotFound'
        '409':
          $ref: '#/components/responses/LastOwnerConflict'
  /servers/{serverId}/invites:
    parameters: [{ $ref: '#/components/parameters/ServerID' }]
    post:
      summary: Create a 24-hour server invite
      operationId: createServerInvite
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Invite created., content: { application/json: { schema: { $ref: '#/components/schemas/Invite' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /servers/{serverId}/invites/{inviteId}/revoke:
    parameters:
      - { $ref: '#/components/parameters/ServerID' }
      - name: inviteId
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    post:
      summary: Revoke a server invite
      operationId: revokeServerInvite
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Invite revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ServerNotFound' }
        '409': { $ref: '#/components/responses/ChannelConflict' }
  /servers/{serverId}/members:
    parameters: [{ $ref: '#/components/parameters/ServerID' }]
    get:
      summary: List server members and their current in-process presence
      operationId: listServerMembers
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Members visible to the caller., content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ServerMember' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/ServerNotFound' }
  /invites/{code}/join:
    post:
      summary: Join a server with an active invite
      operationId: joinServerInvite
      security: [{ bearerAuth: [] }]
      parameters: [{ name: code, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Joined server. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/ServerNotFound' }
  /channels/{channelId}/messages:
    parameters:
      - $ref: '#/components/parameters/ChannelID'
    get:
      summary: List text messages
      operationId: listMessages
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: before, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200': { description: Messages newest first., content: { application/json: { schema: { $ref: '#/components/schemas/MessagePage' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/ChannelNotFound' }
    post:
      summary: Create a text message
      operationId: createMessage
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateMessageRequest' } } }
      responses:
        '201': { description: Message created., content: { application/json: { schema: { $ref: '#/components/schemas/Message' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/ChannelNotFound' }
  /voice/channels/{channelId}/token:
    parameters:
      - $ref: '#/components/parameters/ChannelID'
    post:
      summary: Create a short-lived LiveKit token for a voice channel
      description: >-
        Authorizes a member of the specified voice channel and returns a
        short-lived connection token. The client uses the returned URL and
        token directly with LiveKit; media never passes through this API.
      operationId: createVoiceChannelToken
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Voice session connection details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceJoin'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/VoiceChannelNotFound' }
  /voice/presence:
    put:
      summary: Set or clear the caller's current voice-channel presence
      description: >-
        Updates the caller's in-process voice-channel presence. A channel ID
        marks the caller as connected to that voice channel; null clears the
        current assignment. Changes are delivered to server members through
        the `voice.presence.changed` realtime event.
      operationId: updateVoicePresence
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoicePresenceRequest'
      responses:
        '204': { description: Voice-channel presence updated. }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/VoiceChannelNotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    RegistrationDisabled:
      description: Registration is disabled.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    Conflict:
      description: Username or email already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    Unauthorized:
      description: Credentials or token are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    Forbidden:
      description: The authenticated user does not have the required server role.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    ServerNotFound:
      description: The server does not exist or is not visible to the authenticated user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    LastOwnerConflict:
      description: The final owner of a server cannot leave it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    ChannelNotFound:
      description: The channel does not exist or is not visible to the authenticated user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    ChannelConflict:
      description: A channel with the same name already exists in the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
    VoiceChannelNotFound:
      description: The voice channel does not exist or is not visible to the authenticated user.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorStatus'
  parameters:
    ServerID:
      name: serverId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ChannelID:
      name: channelId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
  schemas:
    LivenessStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          const: alive
          example: alive
    ReadinessStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ready
            - not_ready
          example: ready
    ErrorStatus:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          const: error
          example: error
        code:
          type: string
          description: Stable, program-readable error code.
          example: validation_error
        message:
          type: string
          description: Safe, user-readable error explanation.
          example: Password must contain at least 12 characters.
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 12
    LoginRequest:
      type: object
      required: [identity, password]
      properties:
        identity:
          type: string
          description: Username or email.
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          const: 900
    User:
      type: object
      required: [id, username, email, role, avatar_id]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [owner, member]
        avatar_id:
          type: [string, 'null']
          pattern: '^icon_(0[1-9]|1[0-5])$'
    UpdateAvatarRequest:
      type: object
      required: [avatar_id]
      properties:
        avatar_id:
          type: [string, 'null']
          pattern: '^icon_(0[1-9]|1[0-5])$'
    CreateServerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
    Server:
      type: object
      required: [id, name, description, created_by, membership_role, created_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        created_by:
          type: integer
          format: int64
        membership_role:
          type: string
          enum: [owner, member]
        created_at:
          type: string
          format: date-time
    CreateChannelRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        type:
          type: string
          enum: [text, voice]
          default: text
    Channel:
      type: object
      required: [id, server_id, name, type, position, created_by, created_at]
      properties:
        id:
          type: integer
          format: int64
        server_id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum: [text, voice]
        position:
          type: integer
          minimum: 0
        created_by:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    ServerMember:
      type: object
      required: [id, username, role, avatar_id, online, voice_channel_id]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        role: { type: string, enum: [owner, member] }
        avatar_id: { type: [string, 'null'], pattern: '^icon_(0[1-9]|1[0-5])$' }
        online: { type: boolean }
        voice_channel_id:
          type: [integer, 'null']
          format: int64
          minimum: 1
          description: Current in-process voice channel, or null when not connected.
    Invite:
      type: object
      required: [id, server_id, code, expires_at]
      properties:
        id: { type: integer, format: int64 }
        server_id: { type: integer, format: int64 }
        code: { type: string }
        expires_at: { type: string, format: date-time }
    CreateMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 4000 }
    Message:
      type: object
      required: [id, channel_id, author_id, author_username, author_avatar_id, content, created_at]
      properties:
        id: { type: integer, format: int64 }
        channel_id: { type: integer, format: int64 }
        author_id: { type: integer, format: int64 }
        author_username: { type: string }
        author_avatar_id: { type: [string, 'null'], pattern: '^icon_(0[1-9]|1[0-5])$' }
        content: { type: string }
        created_at: { type: string, format: date-time }
    MessagePage:
      type: object
      required: [messages]
      properties:
        messages: { type: array, items: { $ref: '#/components/schemas/Message' } }
        next_before: { type: [integer, 'null'], format: int64 }
    VoiceJoin:
      type: object
      required: [url, token, room]
      properties:
        url:
          type: string
          format: uri
          description: LiveKit WebSocket URL.
        token:
          type: string
          description: Short-lived LiveKit access token.
        room:
          type: string
          description: Opaque room name derived from the server and voice channel.
    UpdateVoicePresenceRequest:
      type: object
      required: [channel_id]
      properties:
        channel_id:
          type: [integer, 'null']
          format: int64
          minimum: 1
          description: Voice channel to join, or null to clear the current presence.